+++ release_notes.md
@@ -1,2 +1,3 @@
 ## 1.1.0
 * Utility functions and moredash now exported along with bdd-utils
+* migrated entity-resolvers into bdd-utils